requirements.txt
python-telegram-bot
google-cloud-storage
python-dotenv